requirements.txt
Pillow==9.1.1
Django==4.0.5
